Image/Profile Editing Bug Fixes (#157)
* Formatting commands

* fix case where there is no Image field in profile

* fix modal verification and frontend crashing with pfp

* removed unnecessary logger.info

* changed image endpoint to not fail if can't delete

+++ backend/api/views/main.py
@@ -1,7 +1,15 @@
 from os import path
 from flask import Blueprint, request, jsonify
 from bson import ObjectId
-from api.models import db, Education, Video, MentorProfile, AppointmentRequest, Users
+from api.models import (
+    db,
+    Education,
+    Video,
+    MentorProfile,
+    AppointmentRequest,
+    Users,
+    Image,
+)
 from api.core import create_response, serialize_list, logger
 from api.utils.request_utils import (
     MentorForm,
@@ -186,15 +194,17 @@ def uploadImage(id):
         return create_response(status=422, message=msg)
 
     try:
-        if mentor.image.image_hash is True:
+        if mentor.image is True and mentor.image.image_hash is True:
             image_response = imgur_client.delete_image(mentor.image.image_hash)
     except:
-        msg = "Failed to delete image"
-        logger.info(msg)
-        return create_response(status=422, message=msg)
+        logger.info("Failed to delete image but saving new image")
 
-    mentor.image.url = image_response["data"]["link"]
-    mentor.image.image_hash = image_response["data"]["deletehash"]
+    new_image = Image(
+        url=image_response["data"]["link"],
+        image_hash=image_response["data"]["deletehash"],
+    )
+
+    mentor.image = new_image
 
     mentor.save()
     return create_response(status=200, message=f"Success")
